drives: npmx: shell: move gpio code to the separate file
